+++ modules/simple_bucket/versions.tf
@@ -20,12 +20,12 @@ terraform {
 
     google = {
       source  = "hashicorp/google"
-      version = ">= 6.9.0, < 7"
+      version = ">= 6.37.0, < 8"
     }
   }
 
   provider_meta "google" {
-    module_name = "blueprints/terraform/terraform-google-cloud-storage:simple_bucket/v11.0.0"
+    module_name = "blueprints/terraform/terraform-google-cloud-storage:simple_bucket/v12.0.0"
   }
 
 }