+++ versions.tf
@@ -20,7 +20,7 @@ terraform {
 
     google = {
       source  = "hashicorp/google"
-      version = ">= 6.9.0, < 7"
+      version = ">= 6.37.0, < 8"
     }
 
     random = {
@@ -30,7 +30,7 @@ terraform {
   }
 
   provider_meta "google" {
-    module_name = "blueprints/terraform/terraform-google-cloud-storage/v11.0.0"
+    module_name = "blueprints/terraform/terraform-google-cloud-storage/v12.0.0"
   }
 
 }